+++ src/database/model/newsletter.model.js
@@ -1,0 +1,14 @@
+import mongoose from 'mongoose'
+
+const newsSchema = new mongoose.Schema({
+
+    email: {
+        type: String,
+        required: true
+    }
+
+})
+
+
+const News = mongoose.model('News', newsSchema);
+export default News;
